+++ src/frontend/src/features/recording/routes/RecordingDownload.tsx
@@ -57,7 +57,9 @@ const RecordingDownload = () => {
   const isSaved =
     data?.status === RecordingStatus.Saved ||
     data?.status === RecordingStatus.NotificationSucceed ||
-    data?.status === RecordingStatus.FailedToStop
+    data?.status === RecordingStatus.FailedToStop ||
+    data?.status === RecordingStatus.ExternalProcessFailed ||
+    data?.status === RecordingStatus.ExternalProcessSuccessful
 
   const pageTitle = useMemo(() => {
     if (isError) return `${APP_TITLE} - ${t('error.title')}`
@@ -91,6 +93,8 @@ const RecordingDownload = () => {
   }
 
   if (
+    data.status !== RecordingStatus.ExternalProcessFailed &&
+    data.status !== RecordingStatus.ExternalProcessSuccessful &&
     data.status !== RecordingStatus.Saved &&
     data.status !== RecordingStatus.NotificationSucceed &&
     data.status !== RecordingStatus.FailedToStop