+++ src/summary/summary/core/locales/de.py
@@ -30,8 +30,5 @@ Einige Punkte, die wir Ihnen empfehlen zu überprüfen:
         "(externer Link)]({form_link})*\n"
     ),
     hallucination_replacement_text="[Text konnte nicht transkribiert werden]",
-    document_default_title="Transkription",
-    document_title_template=(
-        'Besprechung "{room}" am {room_recording_date} um {room_recording_time}'
-    ),
+    summary_title_template="Zusammenfassung von {title}",
 )